Patch 0688
・x64のWindows環境下において、コンパイル時に「'SOCKET' から 'int' への変換です」の
  警告が発生するのを修正（socket.c）
	reported by Toshiさん
・0687以降、x64でコンパイル時に「'size_t' から 'int' に変換しました。」の
  警告が発生するのを修正（script.c）
	reported by Toshiさん

+++ src/map/script.c
@@ -83,7 +83,7 @@ static int script_pos,script_size;
 
 static struct script_str_data {
 	int type;
-	int str;
+	size_t str;
 	int backpatch;
 	int label;
 	int (*func)(struct script_state *st);
